+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/analytics/DefaultAnalyticsCollector.java
@@ -165,7 +165,6 @@ public class DefaultAnalyticsCollector implements AnalyticsCollector {
 
   // Audio events.
 
-  @SuppressWarnings("deprecation") // Calling deprecated listener method.
   @Override
   public final void onAudioEnabled(DecoderCounters counters) {
     EventTime eventTime = generateReadingMediaPeriodEventTime();
@@ -174,7 +173,6 @@ public class DefaultAnalyticsCollector implements AnalyticsCollector {
         AnalyticsListener.EVENT_AUDIO_ENABLED,
         listener -> {
           listener.onAudioEnabled(eventTime, counters);
-          listener.onDecoderEnabled(eventTime, C.TRACK_TYPE_AUDIO, counters);
         });
   }
 
@@ -190,8 +188,6 @@ public class DefaultAnalyticsCollector implements AnalyticsCollector {
           listener.onAudioDecoderInitialized(eventTime, decoderName, initializationDurationMs);
           listener.onAudioDecoderInitialized(
               eventTime, decoderName, initializedTimestampMs, initializationDurationMs);
-          listener.onDecoderInitialized(
-              eventTime, C.TRACK_TYPE_AUDIO, decoderName, initializationDurationMs);
         });
   }
 
@@ -206,7 +202,6 @@ public class DefaultAnalyticsCollector implements AnalyticsCollector {
         listener -> {
           listener.onAudioInputFormatChanged(eventTime, format);
           listener.onAudioInputFormatChanged(eventTime, format, decoderReuseEvaluation);
-          listener.onDecoderInputFormatChanged(eventTime, C.TRACK_TYPE_AUDIO, format);
         });
   }
 
@@ -240,7 +235,6 @@ public class DefaultAnalyticsCollector implements AnalyticsCollector {
   }
 
   @Override
-  @SuppressWarnings("deprecation") // Calling deprecated listener method.
   public final void onAudioDisabled(DecoderCounters counters) {
     EventTime eventTime = generatePlayingMediaPeriodEventTime();
     sendEvent(
@@ -248,7 +242,6 @@ public class DefaultAnalyticsCollector implements AnalyticsCollector {
         AnalyticsListener.EVENT_AUDIO_DISABLED,
         listener -> {
           listener.onAudioDisabled(eventTime, counters);
-          listener.onDecoderDisabled(eventTime, C.TRACK_TYPE_AUDIO, counters);
         });
   }
 
@@ -282,7 +275,6 @@ public class DefaultAnalyticsCollector implements AnalyticsCollector {
   // Video events.
 
   @Override
-  @SuppressWarnings("deprecation") // Calling deprecated listener method.
   public final void onVideoEnabled(DecoderCounters counters) {
     EventTime eventTime = generateReadingMediaPeriodEventTime();
     sendEvent(
@@ -290,7 +282,6 @@ public class DefaultAnalyticsCollector implements AnalyticsCollector {
         AnalyticsListener.EVENT_VIDEO_ENABLED,
         listener -> {
           listener.onVideoEnabled(eventTime, counters);
-          listener.onDecoderEnabled(eventTime, C.TRACK_TYPE_VIDEO, counters);
         });
   }
 
@@ -306,8 +297,6 @@ public class DefaultAnalyticsCollector implements AnalyticsCollector {
           listener.onVideoDecoderInitialized(eventTime, decoderName, initializationDurationMs);
           listener.onVideoDecoderInitialized(
               eventTime, decoderName, initializedTimestampMs, initializationDurationMs);
-          listener.onDecoderInitialized(
-              eventTime, C.TRACK_TYPE_VIDEO, decoderName, initializationDurationMs);
         });
   }
 
@@ -322,7 +311,6 @@ public class DefaultAnalyticsCollector implements AnalyticsCollector {
         listener -> {
           listener.onVideoInputFormatChanged(eventTime, format);
           listener.onVideoInputFormatChanged(eventTime, format, decoderReuseEvaluation);
-          listener.onDecoderInputFormatChanged(eventTime, C.TRACK_TYPE_VIDEO, format);
         });
   }
 
@@ -345,7 +333,6 @@ public class DefaultAnalyticsCollector implements AnalyticsCollector {
   }
 
   @Override
-  @SuppressWarnings("deprecation") // Calling deprecated listener method.
   public final void onVideoDisabled(DecoderCounters counters) {
     EventTime eventTime = generatePlayingMediaPeriodEventTime();
     sendEvent(
@@ -353,7 +340,6 @@ public class DefaultAnalyticsCollector implements AnalyticsCollector {
         AnalyticsListener.EVENT_VIDEO_DISABLED,
         listener -> {
           listener.onVideoDisabled(eventTime, counters);
-          listener.onDecoderDisabled(eventTime, C.TRACK_TYPE_VIDEO, counters);
         });
   }